SPU2: Pass IOP cycle count at 64bit

+++ pcsx2/SPU2/spu2.h
@@ -81,6 +81,6 @@ void SPU2writeDMA7Mem(u16* pMem, u32 size);
 extern u64 lClocks;
 
 extern void CounterUpdate(u32 DMAICounter);
-extern void TimeUpdate(u32 cClocks);
+extern void TimeUpdate(u64 cClocks);
 extern void SPU2_FastWrite(u32 rmem, u16 value);
